refactor: consolidate assertion options into options.go
